moving searchBar to the separate component
