+++ src/Proifle/Profile.tsx
@@ -107,8 +107,11 @@ const Profile = () => {
                 : [
                     {
                       data: [
-                        { x: 1654003800, y: [0.99, 1, 0.94, 0.95] },
-                        { x: 1654003800, y: [0.99, 1, 0.94, 0.95] },
+                        { x: 1654003800, y: [0.99, 1, 0.95, 0.94] },
+                        { x: 1654003800, y: [0.98, 1.5, 0.96, 0.99] },
+                        { x: 1654003800, y: [0.97, 1.4, 0.97, 0.98] },
+                        { x: 1654003800, y: [0.96, 1.3, 0.98, 0.97] },
+                        { x: 1654003800, y: [0.95, 1.2, 0.99, 0.96] },
                       ],
                     },
                   ]
